# Conflicts:
#	.gitignore
#	Library/MeasureNowPlaying.cpp
#	Library/Util.h

+++ Library/DialogInstall.cpp
@@ -398,7 +398,7 @@ bool DialogInstall::ReadPackage()
 		{
 			const uLong ZIP_UTF8_FLAG = 1 << 11;
 			const DWORD codePage = (ufi.flag & ZIP_UTF8_FLAG) ? CP_UTF8 : CP_ACP;
-			MultiByteToWideChar(codePage, 0, cBuffer, strlen(cBuffer) + 1, buffer, MAX_PATH);
+			MultiByteToWideChar(codePage, 0, cBuffer, (int)strlen(cBuffer) + 1, buffer, MAX_PATH);
 			while (WCHAR* pos = wcschr(buffer, L'/')) *pos = L'\\';
 			return true;
 		}
@@ -709,7 +709,7 @@ bool DialogInstall::InstallPackage()
 		{
 			const uLong ZIP_UTF8_FLAG = 1 << 11;
 			const DWORD codePage = (ufi.flag & ZIP_UTF8_FLAG) ? CP_UTF8 : CP_ACP;
-			MultiByteToWideChar(codePage, 0, cBuffer, strlen(cBuffer) + 1, buffer, MAX_PATH);
+			MultiByteToWideChar(codePage, 0, cBuffer, (int)strlen(cBuffer) + 1, buffer, MAX_PATH);
 			while (WCHAR* pos = wcschr(buffer, L'/')) *pos = L'\\';
 			return true;
 		}
@@ -948,7 +948,7 @@ void DialogInstall::KeepVariables()
 	WCHAR buffer[4];
 	std::wstring currKey, currValue;
 
-	for (int i = 0, isize = m_VariablesFiles.size(); i < isize; ++i)
+	for (size_t i = 0, isize = m_VariablesFiles.size(); i < isize; ++i)
 	{
 		std::wstring fromPath = g_Data.skinsPath + L"@Backup\\";
 		fromPath += m_VariablesFiles[i];
@@ -1298,7 +1298,7 @@ int DialogInstall::IsPluginNewer(const std::wstring& item, const std::wstring& i
 		{
 			const uLong ZIP_UTF8_FLAG = 1 << 11;
 			const DWORD codePage = (ufi.flag & ZIP_UTF8_FLAG) ? CP_UTF8 : CP_ACP;
-			MultiByteToWideChar(codePage, 0, cBuffer, strlen(cBuffer) + 1, buffer, MAX_PATH);
+			MultiByteToWideChar(codePage, 0, cBuffer, (int)strlen(cBuffer) + 1, buffer, MAX_PATH);
 			while (WCHAR* pos = wcschr(buffer, L'/')) *pos = L'\\';
 			return true;
 		}
